Replaced double with float

+++ shaders/mandelbulb_shader.frag
@@ -2,8 +2,8 @@
 
 in vec3 direction;
 
-uniform double step;
-uniform double power;
+uniform float step;
+uniform float power;
 uniform int bail;
 uniform vec3 camera;
 uniform vec3 color;
@@ -19,15 +19,15 @@ bool mandelTest(in vec3 point) {
     return true;
 }
 
-vec3 nextPoint (in vec3 v, in vec3 c, in double power){
+vec3 nextPoint (in vec3 v, in vec3 c, in float power){
 
-	double x = 0;
-	double y = 0;
-	double z = 0;
+	float x = 0;
+	float y = 0;
+	float z = 0;
 	
-	double xx = v.x * v.x;
-	double yy = v.y * v.y;
-	double zz = v.z * v.z;
+	float xx = v.x * v.x;
+	float yy = v.y * v.y;
+	float zz = v.z * v.z;
 	
 	if (equals(power, 1.0))
 		return v;
@@ -40,12 +40,12 @@ vec3 nextPoint (in vec3 v, in vec3 c, in double power){
 		return vec3(x, y, z);
 	}
 	
-    double r = sqrt( xx + yy + zz );
-	double rN = pow( r, power );
-	double nTheta = power * atan( v.y, v.x ) ;
-	double nPhi = power * asin( v.z / r );
+    float r = sqrt( xx + yy + zz );
+	float rN = pow( r, power );
+	float nTheta = power * atan( v.y, v.x ) ;
+	float nPhi = power * asin( v.z / r );
 	
-	double cosNPhi = cos( nPhi);
+	float cosNPhi = cos( nPhi);
 	
 	x = cos(nTheta) * cosNPhi * rN;
 	y = sin(nTheta) * cosNPhi * rN;
@@ -54,24 +54,24 @@ vec3 nextPoint (in vec3 v, in vec3 c, in double power){
 	return vec3(rN*x + c.x, rN*y + c.y, rN*z + c.z);
 }
 
-int equals(in double a, in double b) {
-    double epsilon = 0.000000001;
+int equals(in float a, in double b) {
+    float epsilon = 0.000000001;
     return (abs(a-b) < epsilon);
 }
 
-vec3 rayIntersectsSphere(in vec3 rayPos, in vec3 spherePos, in vec3 rayDir, in double sphereRadius) {
+vec3 rayIntersectsSphere(in vec3 rayPos, in vec3 spherePos, in vec3 rayDir, in float sphereRadius) {
 
     if (length(rayPos-spherePos) <= 2.0) return rayPos;
     vec3 offset = rayPos - spherePos;
     
-    double rSquared = sphereRadius*sphereRadius;
-    double dot = dot(offset, rayDir);
+    float rSquared = sphereRadius*sphereRadius;
+    float dot = dot(offset, rayDir);
 
     if (dot > 0 || dot(offset, offset) < rSquared)
         return vec3(0); // No Collision
 
     vec3 a = offset - dot * rayDir; // plane perpendicular to ray through center
-    double aSquared = a*a;
+    float aSquared = a*a;
 
     if (aSquared > rSquared);
         return vec3(0) // No Collision
